pyproject.toml
[build-system]
requires = ['setuptools']

[project]
name = 'rl-control-template'
version = '0.0.0'
description = ''
authors = [
    {name = 'Parham Panahi', email = 'parham1@ualberta.ca'},
    {name = 'Andy Patterson', email = 'ap3@ualberta.ca'},
    {name = 'Niko Yasui', email = 'yasuiniko@gmail.com'}
]
requires-python = '>=3.11'
readme = 'README.md'
license = {text = 'MIT'}
classifiers = [
    'Programming Language :: Python :: Implementation :: CPython',
    'Programming Language :: Python :: Implementation :: PyPy',
]
dependencies = [
    # custom minatar with jit compiled environments
    'MinAtar @ git+https://github.com/andnp/MinAtar',
    'foragerenv @ git+https://github.com/andnp/forager',
    'gymnasium[atari, box2d]>=1.0.0',
    'optuna',
    'ale_py',
    'pillow',
    'numba>=0.55.0',
    'chex>=0.1.5',
    'jax>=0.2.14',
    'dm-haiku>=0.0.4',
    'optax>=0.0.8',
    'numpy>=1.22.0',
    'PyFixedReps-andnp~=4.0',
    'PyExpUtils-andnp~=7.0',
    'pyrlenvs-andnp~=2.0',
    'ReplayTables-andnp~=6.0',
    'RlGlue-andnp~=1.0',
    'RlEvaluation @ git+https://github.com/rlai-lab/rl-evaluation@ef3a117f7515ba74e7d6ef455e0a2b6a481531b8',
    'PyExpPlotting-andnp',
    'matplotlib',
    'requests>=2.32.3',
    'plantcv~=4.2.1',
    'pandas[excel]',
    'seaborn',
    'supervision',
    'wandb[media]',
    'torch',
    'aiohttp>=3.11.18',
    'aiohttp-retry',
    'scipy<1.16',
]

[dependency-groups]
dev = [
    'pip',
    'ruff',
    'pyright',
    'commitizen',
    'pre-commit',
    'pytest',
    "pytest-asyncio>=1.0.0",
    'pytest-cov'
]

[tool]
[tool.commitizen]
name = 'cz_conventional_commits'
version = '0.0.0'
tag_format = '$version'
version_files = ['pyproject.toml']

[tool.ruff.lint]
select = ['F', 'E', 'W', 'B']
ignore = ['E501', 'E701', 'B023']

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.pyright]
include = ['analysis', 'experiments', 'plots', 'scripts', 'src', 'tests']
venvPath = '.'
venv = '.venv'
typeCheckingMode = 'standard'
useLibraryCodeForTypes = true
